partybot/requirements.txt
python-telegram-bot
openpyxl
parse